requirements.txt
apache-airflow-providers-postgres
apache-airflow-providers-docker
airflow-clickhouse-plugin==1.1.0
airflow-dbt==0.4.0
dbt-core==1.5.8
dbt-greenplum==1.5.0
